refactor: reorganiza pacotes e estrutura de arquivos para imagens
