--- src/main/java/com/projeto/cefom/image/UploadFileResponseDTO.java
@@ -1,9 +1,0 @@
-package com.projeto.cefom.image;
-
-public record UploadFileResponseDTO(
-        String nomeArquivo,
-        String tipoArquivo,
-        long tamanho,
-        String fileDownloadUri
-) {
-}
