+++ src/main/java/com/projeto/cefom/image/dto/UploadFotoAdolescenteResponseDTO.java
@@ -1,0 +1,10 @@
+package com.projeto.cefom.image.dto;
+
+public record UploadFotoAdolescenteResponseDTO(
+        Integer IdFotoAdolescente,
+        String nomeArquivo,
+        String tipoArquivo,
+        long tamanho,
+        String fileDownloadUri
+) {
+}
