+++ src/main/java/com/projeto/cefom/mapper/MatriculaMapper.java
@@ -1,7 +1,7 @@
 package com.projeto.cefom.mapper;
 
 import com.projeto.cefom.dto.response.*;
-import com.projeto.cefom.image.FotoAdolescenteMapper;
+import com.projeto.cefom.image.mapper.FotoAdolescenteMapper;
 import com.projeto.cefom.model.*;
 import com.projeto.cefom.dto.response.AdolescenteEntrevistaResponseDTO;
 import com.projeto.cefom.dto.response.AdolescenteMatriculaEntrevistaResponseDTO;
