+++ src/main/java/com/projeto/cefom/model/Inscricao.java
@@ -1,7 +1,7 @@
 package com.projeto.cefom.model;
 
 import com.projeto.cefom.enums.FinalizacaoInscricao;
-import com.projeto.cefom.image.FotoAdolescente;
+import com.projeto.cefom.image.model.FotoAdolescente;
 import jakarta.persistence.*;
 import java.io.Serializable;
 import java.time.LocalDate;
